refactor(ios): remove unused reporting mode
